+++ src/state/player-teleport.ts
@@ -4,26 +4,22 @@ import { assert } from '../misc/assert'
 import type { Username } from '../net/binary/binary-types'
 import type { MapTpInfo } from '../server/server-types'
 import type { EntityNetid } from '../misc/entity-netid'
+import { runTask } from 'cc-instanceinator/src/inst-util'
 
 declare global {
     interface GlobalStateUpdatePacket {
-        playerTeleport?: Record<
-            Username,
-            {
-                netid: EntityNetid
-                tpInfo: MapTpInfo
-            }
-        >
+        playerTeleport?: Record<Username, TeleportInfoEntry>
     }
 }
 
-let playerTeleports: Record<
-    Username,
-    {
-        netid: EntityNetid
-        tpInfo: MapTpInfo
-    }
-> = {}
+interface TeleportInfoEntry {
+    netid: EntityNetid
+    tpInfo: MapTpInfo
+    fadeIn: number
+    fadeOut: number
+}
+
+let playerTeleports: Record<Username, TeleportInfoEntry> = {}
 
 export const playerTeleportGlobalStateHandler: GlobalStateHandler = {
     get(packet, conn) {
@@ -43,18 +39,16 @@ export const playerTeleportGlobalStateHandler: GlobalStateHandler = {
 
         assert(isRemote(multi.server))
         for (const username in packet.playerTeleport) {
-            const { tpInfo, netid } = packet.playerTeleport[username]
+            const { tpInfo, netid, fadeIn, fadeOut } = packet.playerTeleport[username]
             const client = multi.server.clients.get(username)
             if (!client?.ready) continue
             client.reservedNetid = netid
+            runTask(client.inst, () => ig.game.setTeleportTime(fadeIn, fadeOut))
             client.teleport(tpInfo)
         }
     },
 }
 
-export function notifyRemoteAboutTeleport(username: Username, netid: EntityNetid, tpInfo: MapTpInfo) {
-    playerTeleports[username] = {
-        netid,
-        tpInfo,
-    }
+export function notifyRemoteAboutTeleport(username: Username, entry: TeleportInfoEntry) {
+    playerTeleports[username] = entry
 }